# controller/getdata/resModel/model.go
package resmodel

//import modelscar "park/models/modelsCar"
//
//type Response struct {
//	Message string             `json:"message"`
//	Data    modelscar.CarModel `json:"data"`
//}
//
//type ErrorResponse struct {
//	Error   string `json:"error"`
//	Details string `json:"details"`
//}