# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Fiber API with OAuth2 password flow",
        "title": "Openapi spec",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User Login Data",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "422": {
                        "description": "detail: Validation errors",
                        "schema": {
                            "$ref": "#/definitions/ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Ends the session of a logged-in user by deleting the JWT token cookie.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout User",
                "responses": {
                    "200": {
                        "description": "message: Logout successful",
                        "schema": {
                            "$ref": "#/definitions/MessagedResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieves the current user's username, role, and user ID from the JWT token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "responses": {
                    "200": {
                        "description": "detail: Returns user session information",
                        "schema": {
                            "$ref": "#/definitions/UserSessionExtendedResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "BadRequestResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "BadRequest - invalid request"
                }
            }
        },
...
        "CarParkType": {
            "type": "string",
            "enum": [
                "P1",
                "P2",
                "P3",
                "P4"
            ],
            "x-enum-varnames": [
                "Park1",
                "Park2",
                "Park3",
                "Park4"
            ]
        },
...
        "ChoiceBase": {
            "type": "object",
            "required": [
                "label",
                "value"
            ],
            "properties": {
                "label": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "LoginInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "carPark": {
                    "$ref": "#/definitions/CarParkType"
                },
                "password": {
                    "type": "string",
                    "maxLength": 150
                },
                "username": {
                    "type": "string",
                    "maxLength": 150
                }
            }
        },
        "LoginResponse": {
            "type": "object",
            "required": [
                "access_token",
                "role"
            ],
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/RoleType"
                        }
                    ],
                    "example": "operator"
                }
            }
        },
        "MessagedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "RoleType": {
            "type": "string",
            "enum": [
                "admin",
                "operator",
                "accountant"
            ],
            "x-enum-varnames": [
                "AdminRole",
                "OperatorRole",
                "AccountantRole"
            ]
        },
...
        "UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Unauthorized - Invalid token"
                }
            }
        },
...
        "UserResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "isActive",
                "role",
                "username"
            ],
            "properties": {
                "carPark": {
                    "description": "optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ChoiceBase"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "fullName": {
                    "description": "required in TS, can be empty",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/ChoiceBase"
                },
                "updatedAt": {
                    "description": "optional, can be nil",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "UserSessionExtendedResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "expireAt",
                "id"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expireAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "ipAddress": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "macPassword": {
                    "type": "string",
                    "example": "mac-password"
                },
                "macUsername": {
                    "type": "string",
                    "example": "mac-username"
                },
                "revokedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/UserResponse"
                },
                "userAgent": {
                    "type": "string",
                    "example": "Mozilla/5.0"
                }
            }
        },
...
        "ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Validation error"
                },
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "fieldName": "this field is required"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/v1/auth/login",
            "scopes": {
                "read": "Grants read access",
                "write": "Grants write access"
            }
        }
    }
}